moved weather to service
