+++ src/app/app.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
+import { WeatherService } from './weather/weather.service';
 
 @Component({
   selector: 'app-root',
@@ -11,23 +12,11 @@ export class AppComponent {
   title = 'Dashboard';
   weather: any[] = [];
 
-  constructor(private http: HttpClient) {
-    this.getWeather();
-  }
-
-
-  private getWeather() {
-    console.log('start func');
-    this.http.get('http://weerlive.nl/api/json-data-10min.php?key=51a306b528&locatie=Nijmegen').pipe(map(data => {
-      console.log(data);
-      this.weather.push({
-        key: 'plaats',
-        value: data
-      });
-      // this.weather['temp'] = data['lifeweer'];
-      console.log(this.weather);
-    })).subscribe();
-
+  constructor(
+    private http: HttpClient,
+    private weatherService: WeatherService
+  ) {
+    this.weatherService.refresh();
   }
